requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
aiofiles
passlib
bcrypt
python-jose

